feat: Implement commands for managing attractions, cities, sites, reviews, and sponsors
- Added MarkAttractionAsFeaturedCommand and its handler for marking attractions as featured.
- Enhanced UpdateAttractionCommand to include latitude and longitude, with validation for location.
- Updated AttractionCommandValidators to validate latitude and longitude.
- Modified CreateCityCommand and UpdateCityCommand to accept latitude and longitude instead of GeoLocation.
- Added validation for latitude and longitude in city commands.
- Updated CreateReviewCommand and UpdateReviewCommand to use int for rating and added validation.
- Enhanced DeleteReviewCommand to check user authorization before deleting a review.
- Updated SetReviewStatusCommand to remove the flagging functionality.
- Added CreateSiteCommand and UpdateSiteCommand to accept latitude, longitude, and address details.
- Implemented AddSiteFacilitiesCommand and its DTO without Id.
- Added MarkSiteAsFeaturedCommand for marking sites as featured.
- Updated sponsor commands to use latitude and longitude instead of GeoLocation.
- Enhanced AddSponsorOfferCommand and UpdateSponsorOfferCommand to include date range for offers.
- Added ChangePasswordCommand for user password management.
- Updated LoginCommand to remove RememberMe functionality.

+++ backend/Rafeeq/src/Application/Commands/Reviews/DeleteReviewCommand.cs
@@ -1,12 +1,15 @@
+using Application.Common.Interfaces.Authentication;
 using Domain.Common.Interfaces;
 using Domain.Entities.ReviewAggregate;
+using Domain.Enums;
 
 namespace Application.Commands.Reviews;
 
 public record DeleteReviewCommand(Guid Id) : ICommand;
 
 internal class DeleteReviewCommandHandler(
-    IUnitOfWork unitOfWork) : ICommandHandler<DeleteReviewCommand>
+    IUnitOfWork unitOfWork,
+    IUserContext userContext) : ICommandHandler<DeleteReviewCommand>
 {
     public async Task<Result> HandleAsync(DeleteReviewCommand command, CancellationToken cancellationToken)
     {
@@ -14,7 +17,11 @@ internal class DeleteReviewCommandHandler(
         if (review == null)
             return ReviewErrors.NotFound(command.Id);
 
+        if (!userContext.IsInRole(UserRole.Admin) && review.TouristId != userContext.Id)
+            return Error.Unauthorized("","You are not the owner of this review.");
+
         review.Delete();
+
         await unitOfWork.Reviews.DeleteAsync(review, cancellationToken);
         await unitOfWork.SaveChangesAsync(cancellationToken);
